+++ Info Banks of Moscow/ShowBanksPage.xaml.cs
@@ -24,26 +24,20 @@ namespace Info_Banks_of_Moscow
     {
         public ShowBanksPage()
         {
-            InitializeComponent();
-            //string[] ReadBanks = File.ReadAllLines("Banks.txt");
-
-            //List<Bank> Banks = new List<Bank>();
-
-            BinaryFormatter formatter = new BinaryFormatter();
-
-            using (FileStream fs = new FileStream("Banks.dat", FileMode.OpenOrCreate))
+            try
             {
-                List<Bank> Banks = (List<Bank>)formatter.Deserialize(fs);
-                ShowBanksDataGrid.ItemsSource = Banks;
-                
+                InitializeComponent();
+
+                BinaryFormatter formatter = new BinaryFormatter();
+
+                using (FileStream fs = new FileStream("Banks.dat", FileMode.OpenOrCreate))
+                {
+                    List<Bank> banks = (List<Bank>)formatter.Deserialize(fs);
+                    ShowBanksDataGrid.ItemsSource = banks;
+
+                }
             }
-
-
-            //for (int i = 0; i < ReadBanks.Length; i+=5)
-            //{                
-            //    Bank NewBank = new Bank(ReadBanks[i], ReadBanks[i+1], ReadBanks[i+2], int.Parse(ReadBanks[i+3]), ReadBanks[i+4]);
-            //    Banks.Add(NewBank);
-            //}
+            catch { MessageBox.Show("Что-то пошло не так...", "Ошибка", MessageBoxButton.OK, MessageBoxImage.Error); }
         }
 
         private void FindBankButton_Click(object sender, RoutedEventArgs e)
@@ -59,16 +53,110 @@ namespace Info_Banks_of_Moscow
 
         }
 
+        
         private void DeleteBankButton_Click(object sender, RoutedEventArgs e)
         {
-            DeleteBankPage DeleteBankPage = new DeleteBankPage();
-            NavigationService.Navigate(DeleteBankPage);            
+            try
+            {
+                List<Bank> ableToDeleteBanks = new List<Bank>();
+
+                BinaryFormatter formatter = new BinaryFormatter();
+
+                List<Bank> banks = new List<Bank>();
+
+                using (FileStream fs = new FileStream("Banks.dat", FileMode.OpenOrCreate))
+                {
+                    banks = (List<Bank>)formatter.Deserialize(fs);
+                }
+
+                FileStream fsn = new FileStream("name.txt", FileMode.Open);
+                List<User> users = new List<User>();
+
+                using (StreamReader sr = new StreamReader(fsn))
+                {
+
+                    string str = sr.ReadLine();
+                    string[] loginName = str.Split(new char[] { ' ' }, StringSplitOptions.RemoveEmptyEntries);
+                    User authorisedUser = new User(loginName[1], loginName[0]);
+                    users.Add(authorisedUser);
+                }
+
+
+
+                for (int i = 0; i < banks.Count; i++)
+                {
+                    if (banks[i].GetUserName(banks[i]) == users[0].Login)
+                    {
+                        ableToDeleteBanks.Add(banks[i]);
+                    }
+                }
+                if (ableToDeleteBanks.Count == 0)
+                {
+                    MessageBox.Show("Вы не можете удалить ни одного банка.", "Ошибка", MessageBoxButton.OK, MessageBoxImage.Error);
+                }
+                else
+                {
+                    DeleteBankPage DeleteBankPage = new DeleteBankPage();
+                    NavigationService.Navigate(DeleteBankPage);
+                }
+            }
+            catch { MessageBox.Show("Что-то пошло не так...", "Ошибка", MessageBoxButton.OK, MessageBoxImage.Error); }
         }
 
         private void BackButton_Click(object sender, RoutedEventArgs e)
         {
             ProfilePage ProfilePage = new ProfilePage();
             NavigationService.Navigate(ProfilePage);
+        }
+
+        private void EditBankButton_Click(object sender, RoutedEventArgs e)
+        {
+            try
+            {
+                List<Bank> ableToEditBanks = new List<Bank>();
+
+                BinaryFormatter formatter = new BinaryFormatter();
+
+                List<Bank> banks = new List<Bank>();
+
+                using (FileStream fs = new FileStream("Banks.dat", FileMode.OpenOrCreate))
+                {
+                    banks = (List<Bank>)formatter.Deserialize(fs);
+                }
+
+                FileStream fsn = new FileStream("name.txt", FileMode.Open);
+                List<User> users = new List<User>();
+
+                using (StreamReader sr = new StreamReader(fsn))
+                {
+
+                    string str = sr.ReadLine();
+                    string[] loginName = str.Split(new char[] { ' ' }, StringSplitOptions.RemoveEmptyEntries);
+                    User authorisedUser = new User(loginName[1], loginName[0]);
+                    users.Add(authorisedUser);
+                }
+
+
+
+                for (int i = 0; i < banks.Count; i++)
+                {
+                    if (banks[i].GetUserName(banks[i]) == users[0].Login)
+                    {
+                        ableToEditBanks.Add(banks[i]);
+                    }
+                }
+
+                if (ableToEditBanks.Count == 0)
+                {
+                    MessageBox.Show("Вы не можете изменить ни одного банка.", "Ошибка", MessageBoxButton.OK, MessageBoxImage.Error);
+                }
+                else
+                {
+                    EditBankPage EditBankPage = new EditBankPage();
+                    NavigationService.Navigate(EditBankPage);
+                }
+            }
+            catch { MessageBox.Show("Что-то пошло не так...", "Ошибка", MessageBoxButton.OK, MessageBoxImage.Error); }
         }
     }
 }
